requirements.txt
sentence-transformers==2.6.1
faiss-cpu==1.7.4
numpy
scikit-learn
matplotlib
seaborn
